[test] test spring security

+++ Namukkun/src/main/java/com/pard/namukkun/post/controller/PostController.java
@@ -25,14 +25,22 @@ public class PostController {
 
     @PostMapping("/upload/post")
     @Operation(summary = "게시물 등록", description = "게시물 내용을 입력해서 게시물을 등록합니다.")
-    public ResponseEntity<?> createPost(@RequestBody() PostCreateDTO postCreateDTO) {
+    public ResponseEntity<?> createPost(
+            @SessionAttribute(name = "userid", required = false) Long userId,
+            @RequestBody() PostCreateDTO postCreateDTO
+    ) {
+        if (userId == null) return new ResponseEntity<>(HttpStatus.FORBIDDEN);
         // Post에서 첨부파일을 제외한 데이터는 CreateDTO형태로 받고, 첨부파일은 List형태로 따로 받음
         return postService.createPost(postCreateDTO);
     }
 
     //-----------------------------------
-    @PostMapping("/test")
-    public ResponseEntity<?> test(){
+    @GetMapping("/test")
+    public ResponseEntity<?> test(
+            @SessionAttribute(name = "userid", required = false) Long userId
+    ) {
+        if (userId != 1L)
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
         log.info("test");
         return new ResponseEntity<>(HttpStatus.OK);
     }
@@ -40,13 +48,22 @@ public class PostController {
 
     @PostMapping("/upload/temppost")
     @Operation(summary = "게시물 임시저장", description = "게시물을 임시저장합니다.")
-    public ResponseEntity<?> createTempPost(@RequestBody() PostCreateDTO postCreateDTO) {
+    public ResponseEntity<?> createTempPost(
+            @SessionAttribute(name = "userid", required = false) Long userId,
+            @RequestBody() PostCreateDTO postCreateDTO
+    ) {
+        if (userId == null) return new ResponseEntity<>(HttpStatus.FORBIDDEN);
         return postService.saveTempPost(postCreateDTO);
     }
 
     @PostMapping(value = "/upload/file", consumes = {"multipart/form-data"})
     @Operation(summary = "첨부파일 첨부", description = "첨부파일을 첨부합니다.")
-    public ReturnFileNameDTO uploadFile(@RequestPart("files") List<MultipartFile> files) throws JsonProcessingException {
+    public ReturnFileNameDTO uploadFile(
+            @SessionAttribute(name = "userid", required = false) Long userId,
+            @RequestPart("files") List<MultipartFile> files
+    ) throws JsonProcessingException {
+        if (userId == null) return null;
+
         return postService.uploadAttachment(files);
     }
 
@@ -60,14 +77,20 @@ public class PostController {
 
     @PostMapping(value = "/upload/img", consumes = {"multipart/form-data"})
     @Operation(summary = "이미지 첨부", description = "이미지 이름을 받아서 UUID를 앞에 붙인 이름을 반환합니다.")
-    public ResponseEntity<?> uploadImg(@RequestParam("img") MultipartFile img,
-                                       @SessionAttribute(name = "userinfo", required = false) UserSessionData data){
-        return postService.uploadImg(img,data);
+    public ResponseEntity<?> uploadImg(
+            @RequestParam("img") MultipartFile img,
+            @SessionAttribute(name = "userinfo", required = false) UserSessionData data) {
+        if (data == null) return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        return postService.uploadImg(img, data);
     }
 
     @PostMapping("/delete/file")
     @Operation(summary = "첨부파일 삭제", description = "첨부파일 이름을 받아서 삭제합니다.")
-    public ResponseEntity<?> deleteFile(@RequestParam("fileName") String fileName) {
+    public ResponseEntity<?> deleteFile(
+            @SessionAttribute(name = "userid", required = false) Long userId,
+            @RequestParam("fileName") String fileName
+    ) {
+        if (userId == null) return new ResponseEntity<>(HttpStatus.FORBIDDEN);
         return s3AttachmentService.deleteByName(fileName);
     }
 
@@ -79,8 +102,21 @@ public class PostController {
 
     @GetMapping("/read/update")
     @Operation(summary = "수정할 게시물을 id를 통해서 읽습니다.")
-    public PostUpdateDTO findPostById(@RequestParam("id") Long id) {
-        return postService.findPostByIdUpdateVer(id);
+    public ResponseEntity<?> findPostById(
+            @RequestParam("id") Long id,
+            @SessionAttribute(name = "userid", required = false) Long userId
+
+    ) {
+
+
+        if (userId == null || !postService.getWriterUserId(id).equals(userId))
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+
+
+        PostUpdateDTO dto = postService.findPostByIdUpdateVer(id);
+
+        return new ResponseEntity<>(dto, HttpStatus.OK);
+        //postService.findPostByIdUpdateVer(id);
     }
 
     @GetMapping("/read")
@@ -105,14 +141,19 @@ public class PostController {
 
     @GetMapping("/read/by-Upcount")
     @Operation(summary = "게시물을 추천 높은순으로 정렬해 나열합니다.")
-    public List<PostReadDTO> findByUpCount(){
+    public List<PostReadDTO> findByUpCount() {
         return postService.findByUpCountPost();
     }
 
     @PatchMapping("/update")
     @Operation(summary = "게시물을 수정합니다.", description = "이거 실행하기전에 /post/decreaseUpCount에 가서 파일 첨부먼저 하고 리턴값을" +
             "fileName에 넣어줘야합니다.")
-    public ResponseEntity<?> updatePost(@RequestBody PostUpdateDTO postUpdateDTO, @RequestParam("postId") Long postId) {
+    public ResponseEntity<?> updatePost(
+            @SessionAttribute(name = "userid", required = false) Long userId,
+            @RequestBody PostUpdateDTO postUpdateDTO,
+            @RequestParam("postId") Long postId) {
+        if (userId == null) return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+
         try {
             postService.updatePost(postId, postUpdateDTO);
             return new ResponseEntity<>(HttpStatus.OK);
@@ -124,21 +165,33 @@ public class PostController {
 
     @DeleteMapping("/delete")
     @Operation(summary = "게시물을 삭제합니다.")
-    public ResponseEntity<?> deletePost(@RequestParam("postId") Long postId) {
+    public ResponseEntity<?> deletePost(
+            @SessionAttribute(name = "userid", required = false) Long userId,
+            @RequestParam("postId") Long postId) {
+        if (userId == null || userId.equals(postService.getWriterUserId(postId)))
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
         return postService.deletePost(postId);
     }
 
 
     @PostMapping("/increase/UpCount")
-    @Operation(summary = "게시물 채택" , description = "postid : 채택한 게시물 id, state 추천 상태 true : 채택 / 미채택이면 그냥 없어짐")
-    public ResponseEntity<?> increaeUpCount(@RequestParam("postId") Long postId, @RequestParam("userId") Long userId) {
-        return postService.IncreaseUpCountPost(postId,userId);
+    @Operation(summary = "게시물 채택", description = "postid : 채택한 게시물 id, state 추천 상태 true : 채택 / 미채택이면 그냥 없어짐")
+    public ResponseEntity<?> increaeUpCount(
+            @SessionAttribute(name = "userid", required = false) Long userId,
+            @RequestParam("postId") Long postId
+//            @RequestParam("userId") Long userId
+    ) {
+        if (userId == null) return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        return postService.IncreaseUpCountPost(postId, userId);
     }
 
     @PostMapping("/decrease/UpCount")
     @Operation(summary = "게시물 채택 취소", description = "postid : 채택한 게시물 id, state 추천 상태 true : 채택 / 미채택이면 그냥 없어짐")
-    public ResponseEntity<?> decreaseUpCount(@RequestParam("postId") Long postId, @RequestParam("userId") Long userId) {
-        return postService.decreaseUpCountPost(postId,userId);
+    public ResponseEntity<?> decreaseUpCount(
+            @SessionAttribute(name = "userid", required = false) Long userId,
+            @RequestParam("postId") Long postId) {
+        if (userId == null) return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        return postService.decreaseUpCountPost(postId, userId);
     }
 
 }
